bg/requirements.txt
numpy
torch
torchvision
transformers
fastapi
Pillow
uvicorn[standard]
kornia 
einops 
timm
python-multipart